Float

+++ tests/test_regression.py
@@ -59,7 +59,7 @@ class TestRegression:
             assert torch.mean(torch.abs(T(embeddings1[i]) - T(embeddings0[i]))) <= 1e-6
             assert torch.mean(torch.abs(T(ts1[i]) - T(ts0[i]))) <= 1e-6
         embeddings2, ts2 = torchopenl3.get_audio_embedding(
-            [T(a) for a in audios], srs, batch_size=32, **modelparams
+            [T(a).float() for a in audios], srs, batch_size=32, **modelparams
         )
         for i in range(n):
             """
